requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-filter==22.1
django-import-export==2.9.0
django-ninja==0.19.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==14.2.0
kombu==5.2.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
prompt-toolkit==3.0.31
pydantic==1.10.2
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
typing_extensions==4.3.0
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
